Restore the 1.21.1 Fabric baseline

+++ src/main/java/meteordevelopment/meteorclient/events/render/GetFovEvent.java
@@ -8,9 +8,9 @@ package meteordevelopment.meteorclient.events.render;
 public class GetFovEvent {
     private static final GetFovEvent INSTANCE = new GetFovEvent();
 
-    public float fov;
+    public double fov;
 
-    public static GetFovEvent get(float fov) {
+    public static GetFovEvent get(double fov) {
         INSTANCE.fov = fov;
         return INSTANCE;
     }
